feat: reselect dev tools
