+++ package.json
@@ -78,6 +78,7 @@
     "react-window": "^1.8.6",
     "redux-persist": "^6.0.0",
     "reselect": "^4.1.5",
+    "reselect-tools": "^0.0.7",
     "scrypt-js": "^3.0.1",
     "stream-browserify": "^3.0.0",
     "stream-http": "^3.2.0",
