+++ src/components/AccountAssets/AccountAssets.tsx
@@ -1,10 +1,8 @@
 import { CAIP19 } from '@shapeshiftoss/caip'
 import { Text } from 'components/Text'
 import { selectAssetByCAIP19 } from 'state/slices/assetsSlice/assetsSlice'
-import {
-  AccountSpecifier,
-  selectPortfolioAssetIdsByAccountIdExcludeFeeAsset
-} from 'state/slices/portfolioSlice/portfolioSlice'
+import { AccountSpecifier } from 'state/slices/portfolioSlice/portfolioSlice'
+import { selectPortfolioAssetIdsByAccountIdExcludeFeeAsset } from 'state/slices/portfolioSlice/selectors'
 import { useAppSelector } from 'state/store'
 
 import { Card } from '../Card/Card'
