+++ src/react-app-env.d.ts
@@ -3,3 +3,4 @@
 // Components w/o types
 declare module 'react-qr-reader'
 declare module 'react-middle-ellipsis'
+declare module 'reselect-tools'
